+++ bsn1/CHANGELOG.md
@@ -393,3 +393,4 @@ Nothing is changed.
 ### Added
 
 - Add variant `Anyhow` to enum `Error`
+- Implement `From<anyhow::Error>` for `Error`